// src/generated/enums.ts
export const EChatRoomEnterResponse = {
  k_EChatRoomEnterResponseSuccess: 1,
  k_EChatRoomEnterResponseDoesntExist: 2,
  k_EChatRoomEnterResponseNotAllowed: 3,
  k_EChatRoomEnterResponseFull: 4,
  k_EChatRoomEnterResponseError: 5,
  k_EChatRoomEnterResponseBanned: 6,
  k_EChatRoomEnterResponseLimited: 7,
  k_EChatRoomEnterResponseClanDisabled: 8,
  k_EChatRoomEnterResponseCommunityBan: 9,
  k_EChatRoomEnterResponseMemberBlockedYou: 10,
  k_EChatRoomEnterResponseYouBlockedMember: 11,
  k_EChatRoomEnterResponseRatelimitExceeded: 15,
} as const;
// ...
export const ELobbyComparison = {
  k_ELobbyComparisonEqualToOrLessThan: -2,
  k_ELobbyComparisonLessThan: -1,
  k_ELobbyComparisonEqual: 0,
  k_ELobbyComparisonGreaterThan: 1,
  k_ELobbyComparisonEqualToOrGreaterThan: 2,
  k_ELobbyComparisonNotEqual: 3,
} as const;

export const ELobbyDistanceFilter = {
  k_ELobbyDistanceFilterClose: 0,
  k_ELobbyDistanceFilterDefault: 1,
  k_ELobbyDistanceFilterFar: 2,
  k_ELobbyDistanceFilterWorldwide: 3,
} as const;

export const EChatMemberStateChange = {
  k_EChatMemberStateChangeEntered: 1,
  k_EChatMemberStateChangeLeft: 2,
  k_EChatMemberStateChangeDisconnected: 4,
  k_EChatMemberStateChangeKicked: 8,
  k_EChatMemberStateChangeBanned: 16,
} as const;